Move common components to `src/common/components` We group files by feature, not by type. Consequently, we shouldn't have a `src/components` folder.
