+++ site/src/common/blocks/CallToActionBlock.tsx
@@ -3,7 +3,8 @@ import { PropsWithData, withPreview } from "@comet/cms-site";
 import { CallToActionBlockData } from "@src/blocks.generated";
 import { LinkBlock } from "@src/common/blocks/LinkBlock";
 import { HiddenIfInvalidLink } from "@src/common/helpers/HiddenIfInvalidLink";
-import { Button, ButtonVariant } from "@src/components/common/Button";
+
+import { Button, ButtonVariant } from "../components/Button";
 
 const buttonVariantMap: Record<CallToActionBlockData["variant"], ButtonVariant> = {
     Contained: "contained",
